feat: extract route names to routes.ts (#303)
