+++ packages/client/hmi-client/src/components/sidebar-panel/model-sidebar-panel.vue
@@ -35,7 +35,7 @@ import { onMounted, ref } from 'vue';
 import IconClose32 from '@carbon/icons-vue/es/close/32';
 import { deleteAsset } from '@/services/project';
 import { MODELS } from '@/types/Project';
-import { RouteName } from '@/router';
+import { RouteName } from '@/router/routes';
 import { Model } from '@/types/Model';
 
 const router = useRouter();
