+++ packages/client/hmi-client/src/router/routes.ts
@@ -1,0 +1,11 @@
+export enum RouteName {
+	DatasetRoute = 'dataset',
+	DocumentRoute = 'document',
+	HomeRoute = 'home',
+	ModelRoute = 'model',
+	ProfileRoute = 'profile',
+	ProjectRoute = 'project',
+	ProvenanceRoute = 'provenance',
+	SimulationRoute = 'simulation',
+	SimulationResultRoute = 'simulationResult'
+}
